DefaultTextArea: fix placeholder text alignment
The fix is similar to e3f990999b8738b9d33bfefc1e254149ff873c8a
for TextField. The root cause of the problem is the same - TextArea
doesn't inherit from Control component, so availableWidth and
availableHeight properties are not defined.

But we can use the same way of calculating the width and height of
the PlaceholderText that we use for other styles.

Pick-to: 6.6 6.5
Fixes: QTBUG-120005
Change-Id: I971c8eb5130c8cab8e23cb2b58e089f1490ed1ae
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit c4a1db07c0f833712b66688563efabc9dfce89f2)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>