Fusion style: Set topInset instead of y and height for the background
If the tab button is checked, we want the background to appear raised,
and if not, we want it to appear recessed, with an offset of 2. To do
that, we set y and height in the background delegate of the TabButton,
and binded them to the checked property.
We can do this by setting the topInset instead as it does exactly the
same thing, and reevaluates the height for us, without the extra code.
This is done to not only make the code cleaner, but also as a
workaround to a bug in the evaluation of bindings when setting the
y property in this case.

Fixes: QTBUG-108807
Pick-to: 6.5 6.6 6.7
Change-Id: I7b3639deb1162c6bada81e377a36e090dc398654
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>