Don't call visualFocusChanged() in non-virtual overridden method
visualFocus is a property of QQuickControl and it is potentially
changed when the fcousReason changes.
Now that focusReason is a property of QQuickItem, since
1ca9928856c47108000ffe5f310f2b99740f087b, we override
setFocusReason in QQuickControl and make a call to visualFocusChanged
there. This is not the right approach, since setFocusReason() is not
virtual and the derived class' method will not always be called
(when the method is called via a pointer to the base class) and hence,
the visualFocusChanged will not be emmitted.

To fix, call visualFocusChanged() on the reimplementation of the
virtual focusIn/OutEvent, since that is where we set the focusReason
anyway.

Task-number: QTBUG-117596
Change-Id: I77d061cdb5eb46e2d1cf551b7ec4b81699e09c49
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
(cherry picked from commit 99ce38daa70a1207f507e9fd2d09472567816dc8)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>