tst_qquickitem: Add focusReason() test case
Now that focusReason moved to QQuickItem, we need to add a test for it.
The test is mainly ported from the focusReason() test in qquickcontrol/
tst_focus, but using Item(s) instead of controls.
The test in controls is still kept, as it is interesting to test this
case for compound controls such as SpinBox, ComboBox, etc.

Task-number: QTBUG-117596
Pick-to: 6.7
Change-Id: I087170db490babb8ab19207e7e00aef1a48cd1c5
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>