tst_qquickitem: Add focusReason() test case
Now that focusReason moved to QQuickItem, we need to add a test for it.
The test is mainly ported from the focusReason() test in qquickcontrol/
tst_focus, but using Item(s) instead of controls.
The test in controls is still kept, as it is interesting to test this
case for compound controls such as SpinBox, ComboBox, etc.

Task-number: QTBUG-117596
Change-Id: I087170db490babb8ab19207e7e00aef1a48cd1c5
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
(cherry picked from commit 58a40b1ab2f0607259e604b752d2a9fa9149814d)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>