QQuickSinglePointHandler: check isSinglePointEvent before cast
QQuickSinglePointHandler is not only for QSinglePointEvent, but rather
it's the intermediate base class of any handler that only knows how to
handle one QEventPoint, regardless which device it came from.
The naked cast was quite wrong for touch events, such that the
result returned from buttons() could come from reinterpreted
data of some kind.

Amends a97759a336c597327cb82eebc9f45c793aec32c9 (and then
79cde77f23358adbe57ab8ce08730d2de5bb1288). QQuickPointerEvent
had buttons(), but QPointerEvent doesn't; so the cast became necessary,
but should have been done safely.

f2ba3bd9792500b4d3fcfd23b03098a32641ef4f began checking the buttons
here. In case of a touch event, buttons() always returned NoButton
(regardless whether the touchpoint was being pressed or released);
so in Qt 5, QQuickSinglePointHandler::handlePointerEventImpl() in fact
caused the ungrab and reset() to occur when a touchpoint was released.
We continue doing that: if the event is not a QSinglePointEvent,
buttons() are irrelevant, and we deactivate the handler when handling
the release of the chosen point (as remembered by pointInfo).
But for QSinglePointEvents, checking buttons() is necessary to keep
QTBUG-66360 fixed.

Task-number: QTBUG-66360
Task-number: QTBUG-83980
Fixes: QTBUG-117387
Pick-to: 6.2 6.5 6.6 6.7
Change-Id: I1bc7a72629f5840e660635f2228bf2c0abb7d2bd
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>