Only update the filename text field when the user selects a file
Apparently, it's normal for file dialogs to not update the text field
that represents the filename of the currently selected file, if the
selected file is a directory.

To achieve that behavior, I'm removing the binding on the text property,
and instead call setText() to update the text field when either the user
selects another file in the file dialog list view, or when the
selectedFile is changed externally. But only if the selectedFile is an
actual real file, and not a directory.

Fixes: QTBUG-119917
Pick-to: 6.6 6.5
Change-Id: I8dbf41ba403d09419a2d66130bdad59e66c9d1cf
Reviewed-by: Santhosh Kumar <santhosh.kumar.selvaraj@qt.io>
(cherry picked from commit 8e734b2fc027dc53fca1f344d1ab234a0756ea89)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>