QQuickItem: Respect focusPolicy when delivering events
When delivering events that may potentially transfer focus based on the
focusPolicy of the item, we need to set the focus (if needed) before
letting the item handle the event.
Do this in QQuickItem::event so that the specific event handlers don't
need to deal with setting focus, as well as overrides don't need to
know about those handler's implementations in the base.

Task-number: QTBUG-117596
Pick-to: 6.7
Change-Id: I8223638ce4abf80c212dc63bffdb876254e43442
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>