Make sure updateCurrentTime() is done on with a valid ShaderEffect
The target might change several times, or worse have been destroyed,
while updates are still queued, and since the effect pointer isn't
updated until the postSync() call, the effect pointer can get out of
sync leaving it pointing to the wrong object, or a destroyed one.
To avoid operating on an destroyed or old handle, check if the current
effect is valid and matches the target before applying the uniform
values in updateCurrentTime().

Pick-to: 6.7 6.6 6.5 6.2
Fixes: QTBUG-119363
Change-Id: I5172ccf9405c1d789f197a0cf8cec4c749bd7d28
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>