qmltyperegistrar: Fix QAnyStringViewUtils
The logic for startsWith, endsWith, and split only works if at least one
parameter is encoded in US-ASCII.

To avoid false negatives we would like to make sure that slicing the
first parameter by the second parameter's length results in a substring
that encodes the same number of characters as the second parameter.
This is clearly not always the case if the two parameters are given in
different variable-length encodings.

US-ASCII is a fixed length encoding and all its characters are encoded
with a fixed length in UTF-8 and UTF-16, too. Therefore, if we give the
second parameter in US-ASCII and the result matches, we can be sure it
actually matches. If not, the reason may be:
a, the number of characters covered by the given offset in the first
   parameter is different. In that case we know we have some "high"
   characters that couldn't be encoded in ASCII and cannot match.
b, we have a regular mismatch between two ASCII-encodable character
   sequences.

Both are actual differences between the strings.

Since we have no specific type to hold only US-ASCII strings, we use
QLatin1StringView as the closest type we have.

Change-Id: I8cced40c8e61fc753f7a8ee888513769f99d8085
Reviewed-by: Marc Mutz <marc.mutz@qt.io>