qmlls: fix completion on attached and grouped properties
Clean up the way in which qualified names (like attached properties)
are being completed in bindings and whereever any JSStatement is
expected.

Qualified types are not being supported yet, this is being tracked into
QTBUG-120111.

Also add some QEXPECT_FAIL for the completions that fail because of the
parser.
Currently, we can do:
```
Component.on<request completion here> : {}
```
but we cannot do
```
Component.on<request completion here>
```
as the parser fails because of the missing `: {}`.

Pick-to: 6.7
Task-number: QTBUG-120111
Fixes: QTBUG-120102
Fixes: QTBUG-117442
Change-Id: I6c9ccddcc881aa5253530746286e677407a4326d
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>