FileDialog: set url schema when using the file name text field
When using the text field to choose a file name inside the non-native
file dialog, the selectedFile property would be updated based on the
currentFolder property and what the newly typed filename was.

However, the schema was ignored.

This patch fixes the issue by also setting the schema in
QQuickFileDialogImpl::setFileName(const QString &fileName).

Fixes: QTBUG-120065
Pick-to: 6.7 6.6 6.5
Change-Id: I1860fbbc8209270d0bc6e34a4be6a91bad2253ab
Reviewed-by: Santhosh Kumar <santhosh.kumar.selvaraj@qt.io>