qmlls: adapt parser for attached/grouped properties completion
Add a recovery mode to the parser to accept bindings that are not
completely written out, and create an empty statement for that. This
mode is called enableIncompleteBindings and is only enabled for qmlls.

Also adapt qqmldomastcreator to those empty statements, and qmllsutils.

Fixes: QTBUG-120169
Task-number: QTBUG-92876
Change-Id: Ic24cbb61e3be08368027371e377bf75ce87fafb1
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
(cherry picked from commit 2b008f242150017415bc427d87e84ac81b95e877)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>