QJSValue: document that isUrl only returns true for JS URL objects
It does not return true for a QJSValue that carries a QVariant with a
QUrl. However, on the C++ side the caller will want the URL value anyway
as a QUrl, and in both cases, going through the QVariant works for that.

Fixes: QTBUG-119794
Pick-to: 6.6 6.5
Change-Id: I90fd4a1a27d4dd4758b1060205ce2aeec730eb72
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
(cherry picked from commit 7c59afe361aeb7dee94b9d0093420e6133b8c434)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>