qmlls: fix QmlObject and type annotation completion
Requesting completion on dots on a line before a binding to the default
property was showing incorrect completion items. Fix that by using the
enhanced dom representation of QmlObjects.

This commit enables qmlls to suggest correct completion after `root.`
in cases like
```
x: root.
SomeQualifiedModule.Item {}
```

Also, it enables correct completions for qualified types like
```
SomeQualifiedModule.Item {}
```
It suggest `SomeQualifiedModule.Item` when asking completion for a type,
and only `Item` when asking completions behind a qualifier like
`SomeQualifiedModule.`.

To achieve that, fix the reachableTypes() helper in qqmllsutils to
actually suggest the correct type for qualified type names, now that
there is an adequate Dom representation for qualified type names.

Extend the insideQmlObjectCompletion() helper to assume that it might
actually be completing an incomplete scriptexpression that is before the
actual QmlObject.

Now that a QmlObject contains a DomType::ScriptType for its type, adapt
the code for ScriptType completion to check if it needs completion for
annotated types or for a QmlObject.

This fixes the "binaryExpressionMissingRHSWithDefaultProperty" test,
that was testing scriptexpressions ending in a '.' and followed by a
QmlObject binding to a default property.

Fixes: QTBUG-120111
Task-number: QTBUG-119839
Change-Id: I85ce569355acd320d0823a9c5eb100d8ef0ab098
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
(cherry picked from commit 29d2153f8adeceed56df432ccb775ff28a8e3347)
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>