qmldom: make type a FieldMemberExpression
This commit prepares the Dom to be able to provide correct
completions on binding scripts ending with a `.`.

In addition to saving qualified types as a string, like `QQ.Item` for
example, save them also as a FieldMemberExpression. Do this for
QmlObjects, like `QQ.Item {}` for example, and type annotations, like
`function f(): QQ.Item {}` for example.

This will allow qmlls to suggest completions after `root.` in cases
like
```
x: root.
SomeQualifiedModule.Item {}
```
for example, or after `QQ.` in cases like
```
(x as QQ.Item)
```
for example. The latter magically makes the asCompletion test work
(which was previously QEXPECT_FAIL'd).

Prior to this commit, the Dom did not contain enough information to
know where exactly the completion gets requested inside of
`root.SomeQualifiedModule.Item` and therefore could not suggest any
meaningful suggestion.

Basically reuses a44f21f19462cc79f82080404515c4322b7728ee to model the
nameIdentifiers, a FieldMemberExpression that contains the current type
of the QmlObject.

Adapt existing tests propertyDefinitionBinding and
ignoreNonRelatedTypesForpropertyDefinitionBinding to the changes in the
Dom: the qml object type in a binding, for example `Item` in
`myBinding: Item {}`, is now part of the QmlObject. It was part of the
Binding prior to this commit.

Adapt the qmllsutils method to the change in structure.

Pick-to: 6.7
Task-number: QTBUG-119839
Change-Id: Ie7bc7692f731a01467392dc1dffdf7e67c4d7c46
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>