RegExp: Do on demand JIT compilation
JIT compiling means that we need to obtain executable memory for the
jitted code. That has a non-neglibile cost (in addition to the runtime
cost of compiling the pattern).

Thus, only compile the pattern if we suspect it to be worthwhile:
- when we called match multiple times
- or when the string we match against is quite large

We also assume that the JIT generally works, and thus discard the
bytecode immediately after jitting – if the JIT fails once (either to
compile, or with offset failure), we recompile the bytecode and never
JIT again.

Change-Id: I91e9e10c5ab895af730dbb4a9bc85470dd794d1f
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>