RegExp: Do on demand JIT compilation
JIT compiling means that we need to obtain executable memory for the
jitted code. That has a non-neglibile cost (in addition to the runtime
cost of compiling the pattern).

Thus, only compile the pattern if we suspect it to be worthwhile:
- when we called match multiple times
- or when the string we match against is quite large

We also assume that the JIT generally works, and thus discard the
bytecode immediately before jitting – if the JIT fails once (either to
compile, or with offset failure), we recompile the bytecode and never
JIT this pattern again.

Amends commit 04820be3011395bd5cb13dea8f81c9b9bca4da8f.

Change-Id: I7aed4c0fdd85f1c69bc16cb2caa384aeabd62af2
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>
Reviewed-by: Semih Yavuz <semih.yavuz@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>