QQuickItem::map{To,From}Item: Handle items living in different windows
With child windows now being a possibility, we need to account for
the two items in a map{To,From}Item being in different windows.

This also fixes mapping between items in different top level windows.

The item's scene is defined as the window the item is rendered into,
so nothing changes in that regard. Each window is its own scene;
either another top level scene/window, or a child/sub-scene.

Change-Id: Id2456bf1ab3b913ff9227dd10e90e756ccf33364
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>