Clarify QQuickItem::map{To,From}Item when passed a nullptr
Passing nullptr when mapping to/from an item results in mapping
relative to the scene's coordinate system, and this is also what
the "Scene Coordinates" section of "Concepts - Visual Coordinates
in Qt Quick" documents.

Avoid the ambiguous term "root QML view", which may be mistaken for
the top level window in a child-window scenario.

Pick-to: 6.7 6.6 6.5
Change-Id: Ibd386ec8248f5e398b04248d15c7cf48bd46c090
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>