TextEdit: don't skip nested frames or their blocks in updatePaintNode
I don't fully understand the previous logic behind keeping the same
firstDirtyPos while we iterate all the frames. In case a document
contains a mix of tables and text for example, we may iterate to a block
in the root frame that is just beyond the first table, and set
firstDirtyPos accordingly (which I think is meant to say that everything
up to that character position has been rendered, but it's not true at
that moment in time); and then when we look at the frame corresponding
to the first table, rule out rendering the whole frame or any of its
blocks based on the fact that it comes _before_ firstDirtyPos. The fact
that we got past its position in the root frame should not preclude
rendering the nested frame. So now it looks safer to reset firstDirtyPos
before iterating each nested frame.

Anyway the logic from f513e88403b66c4a5efe4c62c160dfce151efb33 assumed
that we wanted to populate the whole document into the scene graph,
and should just avoid redoing work while keeping the nodes up-to-date;
whereas since 9db23e0e04906cf9ea33e23fa41f34955e5e6fe0 we're trying to
avoid populating nodes too far outside the viewport. That's probably the
bigger optimization: if we are avoiding most of the blocks in large
documents most of the time, redoing work to render the ones that are
in the viewport is not such a big deal. And it seems dangerous to
preemptively skip frames or blocks before we've checked whether they
are in the viewport. So now we visit all frames regardless of
firstDirtyPos, but still skip blocks within the frame that are
beyond the bottom of the viewport. And whenever we do render the
first block that is beyond the viewport, we also enlarge our local
copy of the viewport rectangle to encompass that block's bounds,
to ensure that when we visit nested frames, we will also render
other blocks within the enlarged bounds. Effectively this means
if we render the next block in the root frame immediately past a table,
we will also render all cells in the lower part of that table (between
the top of the viewport and the bottom of the table). That might result
in more SG nodes than we strictly need (outside the viewport bounds);
but at the end, the stored QQuickTextEditPrivate::renderedRegion
must include that one block that was past the table, so we want all
intervening blocks too. QQuickTextEditPrivate::transformChanged() will
not call update() if renderedRegion tells it that the coverage is
complete within the current viewport, and nothing more needs to be
rendered. That was the root of this bug.

The autotest just counts the number of blocks rendered at different
scroll positions in a Flickable. Perhaps it will turn out to be
sensitive to font sizes on various platforms; but the number of
blocks rendered was way too small in case of the bug, so we can
afford a tolerance in the comparison, if needed.

Fixes: QTBUG-118636
Change-Id: Id1f2b393f26e2dd5b198e9d5251e8c33459c3c07
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Doris Verria <doris.verria@qt.io>
(cherry picked from commit db112ff02952437544dad9503b22785b3ff24e99)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>
(cherry picked from commit c30b3e4f96f87b6d350a449d23820ae2f13b3363)