WindowContainer: Decouple contained window from parent on window change
We get the parentWindowChanged() callback when the window the container
item is in is going away, or the container is moved out of the window.

We need to decouple the contained window from the item's window at that
point, as otherwise the item's window will take the contained window
down with it if its destroyed before the container, which doesn't match
the semantics of the window container ultimately managing the window.

Pick-to: 6.7
Change-Id: Ie4640d1e30677a8559a1ba6f19d1eb2aecc87d80
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>