WindowContainer: Don't apply visibility to contained window without parent
A Window with a visual parent will not become visible until it inserted
as a child into another window, as otherwise the Window would be shown
as a top level window.

We handle this by listening for ParentWindowChange events, and then
call applyWindowVisibility(), which applies the Window's visible
state.

The reverse is also the case, when the Window loses the parent window,
in which case we get the same ParentWindowChange callback.

In both these cases, the logic for whether to defer showing the
window or not is contained within applyWindowVisibility(), but the
logic was assuming that the visual parent was another item.

That's not always the case, as a Window can have another Window
as its visual parent. In that case we internally parent the child
window to the contents item of the outer Window, but the visual
parent property still reflects the Window.

To solve this, remove the assumption about the visual parent
always being an item. The deferred visibility logic should
apply in all cases, regardless of what type the visual parent
is.

Pick-to: 6.7
Change-Id: I3f7309ac92274ede72f07f350ce2b1ab65234b83
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>