DOM refactoring. Make DomEnvironment::create return ptr to DomEnv
The core idea inscribed in DOM API is that everything is DomItem,
read-view node.
In most cases RootNode is also a Node. However in the current DOM
design, DomUniverse and DomEnvironment have bigger responsibilities
and wider interface, than just Nodes.
For example, only DomTop types (envs and univ) are allowed to load
files and dependencies.

This commit is a first step to make that separation of concerns
clearer. This commit changes the return type of the
DomEnvironment::create to shared_ptr<DomEnvironment> to make the
interface more explicit.

Change-Id: I94b95fce4c8f96394352126a46dce9122214671a
Task-number: QTBUG-119550
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>