DOM refactoring. Call loadFile on DomUniverse explicitly
As it was mentioned in the previous commit in the chain in the
current design DomItems of variant of DomTop types have wider
interface than just DomItem. Moreover during the loading of the
file, DomEnvironment and DomUniverse are dealing with different
levels of abstraction. The goal of the refactoring is
to make it more explicit.

This commit is a preliminary step towards removing loadFile function
from the DomItem API, preserving its functionality within the
corresponding DomEnvirnment and DomUniverse. This will improve
readability and simplify the mental model.*

*Previously everything was happening through the DomItem API,
the top level function like loadFile was exposed there, and
implementation looked like this:
- "if DomItem is env" and option1 do Load1
- "if DomItem is env" and option2 do Load2
- "if DomItem is env" and option3 do DomItem(Universe)::load
- "if DomItem is universe" then do load3
making it hard to follow and reason about.

This commit removes the last if from the mentioned example
and calls loadFile on the DomUniverse explicitly

Change-Id: I236a33ff2f4edc4c2ab5eeebf99fff9f61bbf3b8
Task-number: QTBUG-119550
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>