DOM refactoring. Make DomItem::loadFile call DomEnv::loadFile
As it was mentioned in the previous commit in the chain, DomUniverse
and DomEnvironment are dealing with different levels of abstraction
during the loading of the files.

This is a preparatory commit for removing DomItem::loadFile. It
introduces DomEnvironment::loadFile containing the DomEnvironment
specific functionality from the DomItem::loadFile and making
DomItem::loadFile to call DomEnvironment::loadFile to verify that
everything is working correctly.

Change-Id: I659ef328ed92cd11719a2e4d9a9151510e401b99
Task-number: QTBUG-119550
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>