qmlls: fix formatting of comments
Fix formatting of comments in qmlls by not formatting a DomItem
that was constructed with the "WithScriptExpression" option.

Instead, create a new Dom representation without the problematic
option.

This is required to fix the CI failure of the 'qmlls: fix QmlObject and
type annotation completion' commit in the relation chain.

Fixes: QTBUG-120484
Pick-to: 6.7
Change-Id: Ic474728455a4938b0e28f4bd4f759d9fc7efef24
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>