DOM refactoring. Remove misleading DomUniverse::options
Assumingly, this option was added when it was planned to have a
threadpool based on the DomUniverse. Existense of the queue, mutexes and
options are clearly handy in that design. The problem is that in the
current approach the queue was only executed inside the DomUniverse,
created with SingleThreaded option.

What's misleading and confusing, that DomUniverse is indeed being used
in the multi-threaded environment (d3c77e13398ba3b288152f2cd3a2934b366ad973)
+ qmlls instantiates a threadpool, where loading of files is happening in
parallel manner.

This commit removes redundant and misleading options.

Change-Id: I8c08a445efde42a7cae291f00f06dce16128e2b6
Task-number: QTBUG-119550
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>