Compiler: Guard against null dereference when ignoring function return
When a function is called, two arrays with the necessary information are
passed to the engine:
  argv:  [return address, prameter 1 address, parameter 2 address, ...]
  types: [return type, parameter 1 type, parameter 2 type, ...]

When the result of the call is ignored, the return type is set to void
and the return address to null.

A check for this null value was missing leading to a null derefence.

Amends: 4f1b9156a48e44cf1f127a4563d0ac69ab436f12

Fixes: QTBUG-120336
Change-Id: I4a21779f3276b0143087b41b0d16c0cd3ba0e7db
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
(cherry picked from commit 22d4a1b7950fa71566c53290a98be3397e910084)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>