DOM refactoring. Remove Queue from DOMUniverse
Assumingly, queue was added to DOMUniverse, when it was planned to
be used in multi-threaded environment. It was a proto model of a
"thread pool". As it was outlined in the previous commit in the
chain, management of threads and threadpools is happening outside the
DomUniverse.
DomUniverse must still be thread-safe, however management of the queue
and parallel execution of tasks is no longer its responsibility.

Hence, this commit removes queue from the DomUniverse.

Change-Id: Icf6a4b15a13bfe556fc0c93c8fdb5b3d5e3f1f17
Task-number: QTBUG-119550
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>